Split DM commands into module
