+++ ledger_bot/process_dm/__init__.py
@@ -1,0 +1,8 @@
+"""Functions to process DM messages."""
+
+from . import command_version, get_dm_channel, is_dm, process_dm
+
+get_dm_channel = get_dm_channel.get_dm_channel
+is_dm = is_dm.is_dm
+process_dm = process_dm.process_dm
+command_version = command_version.command_version
